feat: Tell apart live videos from buffered videos (#464)
* Add flag to `catt cast` to set stream_type manually

Using this flag, I can force catt to see the video I provide as buffered
(fixed-duration video, should be seekable) to a live stream (might not be
seekable, and currently isn't for CATT).

This commit is mostly setting the groundwork for the next commit, where we
autodetect the type of stream using yt-dlp and send it to the PyChromecast
library (which is defaulting to live streams)

* Autodetect if video is stream or fixed-length

Use yt-dlp to detect if the video is a fixed-duration video (buffered)
or a live stream, then send that value to PyChromecast and back to CATT
if and only if it has not been overridden by the --stream-type flag.

* Fixing code and formatting

Added the definition for stream_type in setup_cast() I accidentally deleted

Ran linter with `pre-commit run -a`