Do not guess MIME type of online subtitles
If an URL is specified, the MIME type is read directly from the HTML response,
not guessed from the file termination. With this, we can use URLs that do not
end with .srt, .ttml or .vtt (e.g. URLs with subtitles in YouTube).

To allow this, the file extension is only checked if the file is local.