feat: More control over subtitles (#467)
* Improve handling of a single subtitle file

- TTML is also supported.
- Limits of support are better described.
- Both online and local files are supported.
- A web server is only started if a local file needs to be served.

* guess_mime() used more often, http_server guesses mimetype by itself

* Basic subtitle control in YouTube videos on default controller

Not tested with other platforms.
Only disabling subtitles is compatible with the YouTube controller

* Do not guess MIME type of online subtitles

If an URL is specified, the MIME type is read directly from the HTML response,
not guessed from the file termination. With this, we can use URLs that do not
end with .srt, .ttml or .vtt (e.g. URLs with subtitles in YouTube).

To allow this, the file extension is only checked if the file is local.

* Updating subs_info docstring

Also adding an annotation to the url() method

* Update catt/cli.py

* Fix a KeyError in the new code

yt-dlp cannot always retrieve a name and an extension for a subtitles files.
In these cases, the keys 'ext' and 'name' do not exist, and CATT failed.

Example page: https://iandevlin.github.io/mdn/video-player-with-captions/

---------

Co-authored-by: Stavros Korokithakis <hi@stavros.io>