Features: cast HLS or DASH manifests, override video format
If a user tries to cast from their local computer an HLS or DASH manifest,
and the stream type is not set to LIVE or BUFFERED, CATT will detect it,
issue a warning, and set it to BUFFERED.
CATT will now also send the correct MIME type when such a file is cast.

Support for Smooth Streaming is unchanged for lack of working example files
for testing.

Additionally, CATT silently ignored any format selector provided to yt-dlp.
Currently, if a format selector is provided, CATT will accept this but
warn the user that no compatibility checks will be made. (CATT will not
check if the format is invalid, or simply incompatible with the current
ChromeCast). If no format selector is provided, CATT will silently choose
the best format as usual.

Also StreamInfo.guessed_content_type is now a regular variable and not a
property calculated on the fly (because it depends on ephemeral information
for remote files) and the HTTP server also responds now to HEAD requests, in
order to avoid confusing errors when casting DASH manifests locally.